Preventing the Replay Attack
Finally, in WebhookController, start by querying to see if this event has been handled before. Fetch the EntityManager, and then add $existingLog = ->getRepository('App:StripeEventLog') and call findOneBy() on it to query for stripeEventId set to $eventId

If an  is found, then we don't want to handle this. Just return a new Response() that says "Event previously handled"

If you also want to log a message so that you know when this happens, that's not a bad idea.

But if there is not an existing log, time to process this webhook! Create a new StripeEventLog and pass it . Then, persist and flush just the log

And yea, replay attacks are gone