Adding the Upgrade Button
Now, step two: add an upgrade button that mentions the plan they could switch to. Since we only have 2 plans, it's pretty simple: if they're on the Farmer Brent, we want to allow them to upgrade to the New Zeelander. And if they're on the New Zeelander, we should let them downgrade to the Farmer Brent.

To find the other plan, open SubscriptionHelper and add a new public function called findPlanToChangeTo with a $currentPlanId argument. I'll paste in the logic: it's kind of silly, but it gets the job done. I'm using str_replace instead of something simpler, because in a few minutes, we're going to add yearly plans, and I still want this function to... um... function.

Back to the controller! Add another variable: $otherPlan = null. Then, $otherPlan = ->subscriptionHelper->findPlanToChangeTo() and pass it $currentPlan->getPlanId(). Pass this into the template as an otherPlan variable

There, after the "Active" label, add a button with some classes: a few for styling, one to float right and one - js-change-plan-button - that we'll use in a minute via JavaScript. Make the text: "Change to" and then otherPlan.name

Oh, and add one more attribute: data-plan-name and print otherPlan.name. We'll read that attribute in JavaScript.

Bootstrapping the JavaScript

In fact, let's play with the JavaScript right now: copy the js-change-plan-button class and find the JavaScript block at the top of this file. Use jQuery to locate that element, then on click, add a callback. Start with the always-in-style e.preventDefault()

Start really simple: we'll use a library that I already installed called Sweet Alerts. Call swal() and pass a message Loading Plan Details

Ok, let's see what this Sweet Alerts thing looks like! Refresh that page! Nice! Click the "Change to New Zeelander" link. This is Sweet Alert. It's cute, it's easy, and it'll help us do our job.

Because next, we need to do some serious work: we need to calculate how much we should charge the user to upgrade from the Farmer Brent to the New Zealander, and then show it to the user. That's tricky, because the user is probably in the middle of the month that they've already paid for, so they deserve some credits!

Thankfully, Stripe is going to be a champ and help us out.