Upgrading the Plan in our Database
Ok, the plan is upgraded! Well, in Stripe. But we also need to update the subscription row in our database.

When a user buys a new subscription, we call a method on SubscriptionHelper called addSubscriptionToUser(). We pass it the new \Stripe\Subscription and the User

Then it guarantees that the user has a subscription row in the table with the correct data, like the plan id, subscription id, and $periodEnd date.

Now, the only thing we need to update right now is the plan ID: both the subscription ID and period end haven't changed. But that's ok: we can still safely reuse this method.

In ProfileController, add $this->subscriptionHelper->addSubscriptionToUser() passing it $stripeSubscription and $this->getUser()

And that's everything. At the bottom... well, we don't really need to return anything to our JSON. So just return a new Response() with null as the content and a 204 status code

This doesn't do anything special: 204 simply means that the operation was successful, but the server has nothing it wishes to say back.

Executing the Upgrade in the UI

Copy the route name, then head to the template to make this work.

First, find the button, copy the data-preview-url attribute, and paste it. Name the new one data-change-url and update the route name. Above in the JavaScript, set a new changeUrl variable to .data('change-url'). Then, scroll down to the bottom: this callback function will be executed if the user clicks the "Ok" button to confirm the change. Make the AJAX call here: set the url to changeUrl, the method to POST, and attach one more success function. Inside that, call Sweet Alert to tell the user that the plan was changed! Let's also add some code to reload the page after everything

OK! Let's do this! Refresh the page! Click to change to the "New Zeelander". $99.88 - that looks right, now press "Ok". And ... cool! I think it worked! When the page reloads, our plan is the "New Zeelander" and we can downgrade to the "Farmer Brent".

In the Stripe dashboard, open payments, click the one for $99.88, and open its Invoice. Oh, it's a thing of beauty: this has the two line items for the change.

If you check out the customer, their top subscription is now to the New Zealander plan.

So we're good. Except for one last edge-case.