Submitting the Coupon Form
So let's move to phase two: when we hit Add, this should submit to a new endpoint that validates the code in Stripe and attaches it to our user's cart.

To create the new endpoint, open OrderController. Near the bottom add a new public function addCouponAction() with @route("/checkout/coupon"). Name it order_add_coupon. And to be extra-hipster, add methods{"POST"} to guarantee that you can only POST to this.
Cool! Copy the route name, then find the coupon form in the checkout template. Update the form's action: add path() and then paste the route name. Next, we'll read the submitted code and check with Stripe to make sure it's real, and not just someone trying to guess clever coupon codes. Come on, we've all tried it before.

The coupon form will submit its code field right in addCouponAction on OrderController
To fetch that POST parameter, add the Request object as an argument. Then add $code = $request->request->get('code'). And in case some curious user submits while the form is empty, send back a validation message with $this->addFlash('error', 'Missing coupon code'). Redirect to the checkout page

Fetching the Coupon Information

Great! At this point, all we need to do is talk to Stripe and ask them:

Hey Stripe! Is this a valid coupon code in your system?. Oh, and if it is, ow much is it for?

Since Coupon is just an object in Stripe's API, we can fetch it like anything else. Booya!

As usual, add the API call in StripeClient. At the bottom, create a new public function called findCoupon() with a $code argument. Then, return \Stripe\Coupon::retrieve() and pass it the  string, which is the Coupon's primary key in Stripe's API. Back in OrderController, add $stripeCoupon = $this->stripeClient and then call ->findCoupon($code)

If the code is invalid, Stripe will throw an exception. We'll handle that in a few minutes. But just for now, let's dump() and die to see what it looks like.

Ok, refresh, hit "I have a coupon code," fill in our CHEAP_SHEEP code, and submit!

There it is! In the _values section where the data hides, the coupon has an id, it shows the amount_off in cents and has a few other things, like duration, in case you want to create coupons that are recurring and need to tell the user that this will be applied multiple times.