Showing the Code on Checkout
But no worries, the method below getTotal() - getTotalWithDiscount() - subtracts the coupon code

So back in the template, use cart.totalWithDiscount
Ah, now it shows $49.

But, it'll be even clearer if we display the discount in the table. At the bottom of that table, add a new if statement: if cart.couponCode and an endif. Then, copy the subscription block from above, paste it here, and change the first variable to cart.couponCode and the second to cart.couponCodeValue without the / month, unless you want to make all your coupons recurring. Oh, and add "Coupon" in front of the code

This time, the whole page makes sense! $99 - $50 = $49. It's a miracle!

Now for the easy step: apply the coupon to the user's order at checkout... ya know, so that they actually save $50.