Go create a huge coupon - like for 00. Call this one SUPER_SHEEP!
Ok, this coupon is awesome - so let's add it to an order.

Perfect! As you can see, the cart is already smart enough to return the total as $0, instead of a negative number. Way to go cart!

But, uh, what's this checkout form still doing over here? Why should I need to enter my credit card info? This order is free!

Look, how you handle free orders is up to you. If you still want to require a credit card, you can do that. The customer won't be charged, but the card will be on file for renewals.

But that's not for me: I want to make it as easy as possible for this user to checkout free.
Add all methods necessary to empty cart