Hide the Checkout Form
The first step, is to hide this checkout form! In checkout.html.twig, find the _cardForm.html.twig include, and wrap it in if cart.totalWithDiscount > 0. Else, create a really simple form that submits right back to this URL. Give it method="POST" and a submit button that invites the user to "Checkout for Free"

If you refresh now, boom! Checkout form gone.

Handling Free Checkout

But that's not quite everything we need to do. Thanks to the discount, Stripe will already know that it doesn't need to charge the user, and so, the customer doesn't need to have a card. But, the funny thing is, up until now, our OrderController logic is expecting a stripeToken to always be submitted:
169 lines  src/Controller/OrderController.php
class OrderController extends BaseController
{
    public function checkoutAction(Request request)
    {
        if ($request->isMethod('POST')) {
            $token = $request->request->get('stripeToken');
        }
    }
}
Remember that's the token that we get back from Stripe, after submitting the credit card information to them. We then pass that to chargeCustomer() and attach it to the Customer:

 169 lines  src/Controller/OrderController.php
class OrderController extends BaseController
{
    private function chargeCustomer($token)
    {
        if (! $user->getStripeCustomerId()) {
            $stripeCustomer = $stripeClient->createCustomer($user, $token);
        } else {
            $stripeCustomer = $stripeClient->updateCustomerCard($user, $token);
        }
    }
}

But now, our code needs to be smart enough to not try to attach the token to the Customer for free orders.